+++ src/settings.py
@@ -92,6 +92,22 @@ PLUGIN_ACTION_INSTALL = 'install'
 PLUGIN_ACTION_UNINSTALL = 'uninstall'
 PLUGIN_ACTION_NOTHING = 'nothing'
 
+""" Tables in which configuration is stored, with 'auto gen id' fields and 'unique field'
+(others than only auto-gen field). Those tables must be sorted to satisfy foreign keys.
+Those are the 'short names' of the tables. We will need to add WordPress table prefix to
+have complete name. """
+WP_PLUGIN_CONFIG_TABLES = {'postmeta': ['meta_id', None],
+                           'options': ['option_id', 'option_name'],
+                           'terms': ['term_id', None],
+                           'termmeta': ['meta_id', None],
+                           'term_taxonomy': ['term_taxonomy_id', None],
+                           'term_relationships': [None, ['object_id', 'term_taxonomy_id']]}
+
+""" Relation between configuration tables. There are no explicit relation between tables in DB but there are
+relation coded in WP. """
+WP_PLUGIN_TABLES_RELATIONS = {'termmeta': {'term_id': 'terms'},
+                              'term_taxonomy': {'term_id': 'terms'},
+                              'term_relationships': {'term_taxonomy_id': 'term_taxonomy'}}
 
 JAHIA_USER = Utils.get_optional_env("JAHIA_USER", "admin")
 JAHIA_HOST = Utils.get_optional_env("JAHIA_HOST", "localhost")
