+++ src/wordpress/__init__.py
@@ -1,9 +1,11 @@
 from .models import WPException, WPSite, WPUser
 from .config import WPConfig
 from .themes import WPThemeConfig
-from .plugins import WPPluginConfig, WPPluginConfigExtractor, WPPluginList
 from .generator import WPGenerator
 from .backup import WPBackup
+from .plugins.models import WPPluginList, WPPluginConfigInfos
+from .plugins.config import WPPluginConfig
+from .plugins.manager import WPPluginConfigExtractor, WPPluginConfigRestore
 
 
 __all__ = [
@@ -13,8 +15,10 @@ __all__ = [
     WPGenerator,
     WPBackup,
     WPConfig,
+    WPPluginList,
+    WPPluginConfigInfos,
     WPPluginConfig,
     WPPluginConfigExtractor,
+    WPPluginConfigRestore,
     WPThemeConfig,
-    WPPluginList,
 ]
