+++ src/wordpress/generator.py
@@ -13,7 +13,8 @@ from veritas.validators import validate_string, validate_openshift_env, validate
 from .models import WPSite, WPUser
 from .config import WPConfig
 from .themes import WPThemeConfig
-from .plugins import WPPluginList, WPPluginConfig
+from .plugins.models import WPPluginList
+from .plugins.config import WPPluginConfig
 
 
 class WPGenerator:
