+++ src/wordpress/plugins/polylang.py
@@ -1,3 +1,5 @@
+from .config import WPPluginConfig
+
 
 class WPPolylang(WPPluginConfig):
     pass
